requirements.txt
Django>=3.0,<4.0
mysqlclient
mysql-connector-python
Pillow
xlwt
